fix: fix bug when remove an item that already exists in the cart, but with sizes different
The application considers that two similiar products with different sizes aren't
the equals. This way, when user removes, an example, one product with size P,
but there is also the same product with the size G. So, the application should
removes the product with P, but it wasn't doing that.

It was necessary to create another logic to change the product size, because
the logic used it was kind of wrong.